# backend/internal/apidocs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Habitus Partner API",
    "version": "2026-09-06",
    "summary": "Подбор жилья по запросу словами — как встраиваемый сервис.",
    "description": "Habitus разбирает запрос на естественном языке, подбирает объекты и объясняет каждый выбор фактами о районе: сколько ребёнку идти до школы, куда выходят окна, что вокруг по вечерам.\n\nPartner API отдаёт то же ядро, что работает в самом продукте: поиск, разбор запроса, досье объекта с проверяемыми утверждениями, гео-слои города, управление собственной витриной и заявки покупателей.\n\n**Главное правило сервиса — не выдумывать факты о городе.** Если замера нет, поле придёт как null, а блок разбора деградирует или исчезнет. Нулей вместо отсутствующих измерений здесь не бывает; модельные оценки всегда помечены как proxy.",
    "termsOfService": "https://habitus.city/terms",
    "contact": {
      "name": "Поддержка интеграций",
      "email": "api@habitus.city",
      "url": "https://habitus.city/docs"
    },
    "license": {
      "name": "Проприетарная лицензия",
      "url": "https://habitus.city/terms"
    }
  },
  "servers": [
    {
      "url": "https://api.habitus.city",
      "description": "Продакшен"
    },
    {
      "url": "http://localhost:8080",
      "description": "Локальная разработка (docker compose up)"
    }
  ],
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/partner/v1/searches/{search_id}": {
      "get": {
        "operationId": "getSearch",
        "tags": [
          "Поиск"
        ],
        "summary": "Метаданные поиска",
        "description": "Разбор запроса, ослабленные требования, недоступные слои и объяснение — без самой выдачи.",
        "parameters": [
          {
            "$ref": "#/components/parameters/SearchID"
          }
        ],
        "responses": {
          "200": {
            "description": "Поиск.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Search"
                }
              }
            }
          },
          "401": {
            "description": "Ключ не передан, не распознан, отозван или истёк.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "invalid_api_key": {
                    "value": {
                      "error": {
                        "type": "authentication_error",
                        "code": "invalid_api_key",
                        "message": "Ключ API не распознан. Проверьте заголовок Authorization: Bearer hab_live_…",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Ключу не хватает права, либо доступ партнёра приостановлен.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "insufficient_scope": {
                    "value": {
                      "error": {
                        "type": "permission_error",
                        "code": "insufficient_scope",
                        "message": "Ключу не хватает права search:read",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors",
                        "param": "scope"
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Превышен лимит. Заголовок Retry-After говорит, через сколько секунд повторять.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "type": "rate_limit_error",
                        "code": "rate_limit_exceeded",
                        "message": "Превышен лимит запросов (120 в минуту). Повторите через 12 с",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Поиск не найден или уже удалён по сроку хранения",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "search_not_found": {
                    "value": {
                      "error": {
                        "type": "invalid_request_error",
                        "code": "search_not_found",
                        "message": "Поиск не найден или уже удалён по сроку хранения",
                        "request_id": "01J8Z5C3K9QW2Y6R7T8V9XA0BC",
                        "doc_url": "https://api.habitus.city/docs#errors"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "hab_live_…",
        "description": "Ключ передаётся заголовком `Authorization: Bearer hab_live_…`. Секрет показывается один раз при выпуске и в системе не хранится — восстановить его нельзя, только выпустить новый.\n\nПрава ключа задаются скоупами: `search:read`, `geo:read`, `listings:read`, `listings:write`, `leads:read`, `webhooks:write`."
      }
    },
    "parameters": {
...
      "SearchID": {
        "name": "search_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Идентификатор поиска из POST /search."
      },
...
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "description": "Единый конверт ошибки. Один и тот же для всех ручек и всех кодов ответа.",
        "properties": {
          "error": {
            "type": "object",
            "description": "",
            "properties": {
              "type": {
                "type": "string",
                "description": "Класс ошибки: authentication_error, permission_error, invalid_request_error, rate_limit_error, conflict_error, api_error. По нему клиент решает — чинить запрос, повторить позже или писать нам.",
                "enum": [
                  "authentication_error",
                  "permission_error",
                  "invalid_request_error",
                  "rate_limit_error",
                  "conflict_error",
                  "api_error"
                ]
              },
              "code": {
                "type": "string",
                "description": "Машинный код конкретной ошибки, например invalid_api_key или idempotency_key_reuse."
              },
              "message": {
                "type": "string",
                "description": "Человеческое описание. Пригодно для показа оператору, но ветвиться нужно по code."
              },
              "param": {
                "type": "string",
                "description": "Поле запроса, из-за которого отказ. Отсутствует, если отказ не про конкретное поле."
              },
              "cause": {
                "type": "string",
                "description": "Техническая улика: стадия, статус вышестоящего сервиса, тайминги. Появляется только когда она есть."
              },
              "hint": {
                "type": "string",
                "description": "Что с этим делать. Появляется только когда есть что посоветовать."
              },
              "request_id": {
                "type": "string",
                "description": "Идентификатор запроса. Тот же, что в заголовке X-Request-Id — назовите его в обращении в поддержку."
              },
              "doc_url": {
                "type": "string",
                "description": "Ссылка на раздел документации об ошибках."
              }
            },
            "required": [
              "type",
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Search": {
        "type": "object",
        "description": "Поиск: разбор запроса и то, что с ним пришлось сделать.",
        "properties": {
          "object": {
            "type": "string",
            "description": "",
            "enum": [
              "search"
            ]
          },
          "id": {
            "type": "string",
            "description": "Идентификатор поиска. Нужен для выдачи, досье и вопросов по объекту.",
            "format": "uuid"
          },
          "query": {
            "type": "string",
            "description": "Исходный запрос."
          },
          "city": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "description": "",
            "format": "date-time"
          },
          "parsed": {
            "type": "object",
            "description": "Что система поняла из запроса: диапазон цены, комнаты, площадь, гео-требования, маршруты членов семьи. Структура повторяет разбор поискового ядра."
          },
          "relaxed": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Требования, которые пришлось ослабить, чтобы выдача не оказалась пустой. Пустой массив — ничего не ослабляли."
          },
          "degraded": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Слои данных, недоступные в момент запроса. Их отсутствие уже учтено в оценках, но знать о нём нужно."
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Честные примечания о покрытии данных. Например: ориентация окон известна у малой доли объявлений, поэтому учтена как предпочтение, а не как фильтр."
          },
          "explanation": {
            "type": "string",
            "description": "Текстовое объяснение выдачи. Пустая строка, если explain не запрашивали."
          },
          "data_freshness": {
            "type": [
              "string",
              "null"
            ],
            "description": "Дата актуальности данных. null — замера нет."
          },
          "area_label": {
            "type": [
              "string",
              "null"
            ],
            "description": "Название зоны, в которой шёл поиск, если она была определена."
          },
          "area_geojson": {
            "description": "Граница зоны поиска в GeoJSON. null, если зоны не было.",
            "type": [
              "object",
              "null"
            ]
          },
          "total": {
            "type": "integer",
            "description": "Сколько объектов сохранено под этим поиском."
          }
        }
      },
...
    }
  },
...
}